requirements.txt
django-crispy-forms==1.8.1
